Move responses to model
