+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/model/NewSimulationResponse.kt
@@ -1,0 +1,3 @@
+package app.urbanflo.urbanflosumoserver.model
+
+data class NewSimulationResponse(val id: String)
