+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/model/SimulationInfo.kt
@@ -1,0 +1,3 @@
+package app.urbanflo.urbanflosumoserver.model
+
+data class SimulationInfo(val id: String)
