+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/storage/StorageService.kt
@@ -1,7 +1,7 @@
 package app.urbanflo.urbanflosumoserver.storage
 
 import app.urbanflo.urbanflosumoserver.SimulationInstance
-import app.urbanflo.urbanflosumoserver.responses.SimulationInfo
+import app.urbanflo.urbanflosumoserver.model.SimulationInfo
 import org.springframework.core.io.Resource
 import org.springframework.web.multipart.MultipartFile
 import java.util.*
